backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
aiofiles==23.2.0
aiohttp==3.8.6
pandas==2.1.3
numpy==1.26.2
yfinance==0.2.64
alpha-vantage==2.3.1
beautifulsoup4==4.12.2
playwright==1.40.0
feedparser==6.0.10
google-generativeai==0.3.2
langchain==0.0.350
langchain-community==0.0.3
python-dotenv==1.0.0
celery[redis]==5.3.4
redis==4.6.0
ta==0.10.2
schedule==1.2.0
loguru==0.7.2
pytz==2023.3 
readability-lxml==0.8.1
lxml[html_clean]==4.9.3
newspaper3k==0.2.8
python-dateutil==2.8.2 